also move config objects to testFixtures
