+++ airbyte-cdk/bulk/core/load/src/integrationTest/kotlin/io/airbyte/cdk/load/mock_integration_test/MockDestinationChecker.kt
@@ -6,6 +6,7 @@ package io.airbyte.cdk.load.mock_integration_test
 
 import io.airbyte.cdk.load.check.DestinationChecker
 import io.airbyte.cdk.load.test.mock.MockDestinationBackend.MOCK_TEST_MICRONAUT_ENVIRONMENT
+import io.airbyte.cdk.load.test.mock.MockDestinationConfiguration
 import io.micronaut.context.annotation.Requires
 import javax.inject.Singleton
